ref(test): Move mock endpoint code to module
Done in preparation for a refactor of how we construct mock endpoints
