+++ tests/integration/test_utils/mod.rs
@@ -1,3 +1,6 @@
 //! A collection of utilities for integration tests.
 
 pub mod env;
+mod mocking;
+
+pub use mocking::{mock_endpoint, EndpointOptions};
